moved hackerNews util to lib folder
