+++ lib/hackerNews.ts
